bugfix and testing
